Change float by double
git-svn-id: file:///home/roberto/Git/svn@10119 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 593808bbfd1e94f640d2238b29d563ff3aad99e5

+++ trunk/xmipp/libraries/data/numerical_recipes.cpp
@@ -9625,7 +9625,7 @@ double gammp(double a, double x)
 void choldc(double *a, int n, double *p)
 {
     int i, j, k;
-    float sum;
+    double sum;
 
     for (i = 1;i <= n;i++)
     {
@@ -9648,7 +9648,7 @@ void choldc(double *a, int n, double *p)
 void cholsl(double *a, int n, double *p, double *b, double *x)
 {
     int i, k;
-    float sum;
+    double sum;
 
     for (i = 1;i <= n;i++)
     {
